move compress and thumbnail function to utils
